Apply product filtering, improve formatting and ts types and validations

+++ src/app/components/registration/registration.component.ts
@@ -1,6 +1,19 @@
 import { Component } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { AbstractControl, FormControl, FormGroup, ValidationErrors, Validator, ValidatorFn, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { trimValidator } from 'src/app/trim-validator';
+import { User } from 'src/app/user';
+
+function validationSequence(validators: ValidatorFn[]): ValidatorFn {
+  return (control: AbstractControl): ValidationErrors | null => {
+    for (const validator of validators) {
+      if (validator(control)) {
+        return validator(control);
+      }
+    }
+    return null;
+  }
+}
 
 @Component({
   selector: 'app-registration',
@@ -9,19 +22,23 @@ import { Router } from '@angular/router';
 })
 export class RegistrationComponent {
 
-  constructor(private router : Router){}
+  constructor(private router: Router) { }
 
   registrationForm = new FormGroup({
-      name:new FormControl('',Validators.required),
-      email:new FormControl('',[Validators.required,Validators.email]),
-      password:new FormControl('',[Validators.required,Validators.minLength(6)])
-    })
+    name: new FormControl('', validationSequence([Validators.required, Validators.minLength(3), trimValidator()])),
+    email: new FormControl('', validationSequence([Validators.required, Validators.email, trimValidator()])),
+    password: new FormControl('', validationSequence([Validators.required, Validators.minLength(6), trimValidator()]))
+  })
 
-    onSubmit(){
-      const users = JSON.parse(localStorage.getItem('users') || '[]');
-      users.push({...this.registrationForm.value,isLoggedIn:false});
-      localStorage.setItem('users',JSON.stringify(users));
-      this.registrationForm.reset()
-      this.router.navigate(['/'])
+  onSubmit() {
+    const users = JSON.parse(localStorage.getItem('users') || '[]');
+    const existEmail = users.some((user: User) => user.email === this.registrationForm.get('email')?.value)
+    if (existEmail) {
+      return alert('user already exist!!')
     }
+    users.push({ ...this.registrationForm.value, isLoggedIn: false });
+    localStorage.setItem('users', JSON.stringify(users));
+    this.registrationForm.reset()
+    this.router.navigate(['/'])
+  }
 }
